+++ .gitignore
@@ -26,8 +26,8 @@
 .env.production.local
 .eslintcache
 .idea
-cypress/videos
-cypress/screenshots
+e2e/videos
+e2e/screenshots
 
 npm-debug.log*
 yarn-debug.log*
