--- cypress/config/dev.json
@@ -1,13 +1,0 @@
-{
-  "baseUrl": "https://dev.tgrade.finance/",
-  "defaultCommandTimeout": 25000,
-  "viewportWidth": 1920,
-  "viewportHeight": 1080,
-  "watchForFileChanges": false,
-  "video": true,
-  "videoCompression": false,
-  "screenshotOnRunFailure": true,
-  "retries": {
-    "runMode": 0
-  }
-}
