--- cypress/config/local.json
@@ -1,12 +1,0 @@
-{
-  "baseUrl": "http://localhost:3000/",
-  "defaultCommandTimeout": 80000,
-  "viewportWidth": 1920,
-  "viewportHeight": 1080,
-  "watchForFileChanges": false,
-  "video": true,
-  "screenshotOnRunFailure": true,
-  "retries": {
-    "runMode": 0
-  }
-}
