--- cypress/plugins/index.ts
@@ -1,48 +1,0 @@
-/// <reference types="cypress" />
-// ***********************************************************
-// This example plugins/index.ts can be used to load plugins
-//
-// You can change the location of this file or turn off loading
-// the plugins file with the 'pluginsFile' configuration option.
-//
-// You can read more here:
-// https://on.cypress.io/plugins-guide
-// ***********************************************************
-
-// This function is called when a project is opened or re-opened (e.g. due to
-// the project's config changing)
-
-/**
- * @type {Cypress.PluginConfig}
- */
-// eslint-disable-next-line no-unused-vars
-const browserify = require("@cypress/browserify-preprocessor");
-const cucumber = require("cypress-cucumber-preprocessor").default;
-const resolve = require("resolve");
-const fs = require("fs-extra");
-const path = require("path");
-
-function getConfigurationByFile(file) {
-  const pathToConfigFile = path.resolve("..", "config", `${file}.json`);
-  return fs.readJson(pathToConfigFile);
-}
-
-module.exports = (on, config) => {
-  const options = {
-    ...browserify.defaultOptions,
-    typescript: resolve.sync("typescript", { baseDir: config.projectRoot }),
-  };
-  on("file:preprocessor", cucumber(options));
-
-  on("task", {
-    log(message) {
-      console.log(message);
-
-      return null;
-    },
-  });
-  // accept a configFile value or use development by default
-  const file = config.env.configFile || "stage";
-
-  return getConfigurationByFile(file);
-};
