+++ package.json
@@ -22,7 +22,7 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.0.2",
     "@types/ledgerhq__hw-transport-webusb": "^4.70.1",
-    "@types/node": "^16.11.6",
+    "@types/node": "^17.0.42",
     "@types/react": "^17.0.33",
     "@types/react-burger-menu": "^2.8.2",
     "@types/react-dom": "^17.0.10",
@@ -62,11 +62,13 @@
   },
   "devDependencies": {
     "@cypress/browserify-preprocessor": "^3.0.2",
+    "@cypress/webpack-preprocessor": "^5.12.0",
     "@testing-library/cypress": "^8.0.2",
     "axios": "^0.26.1",
-    "cypress": "^10.0.2",
+    "cypress": "^10.1.0",
     "cypress-cucumber-preprocessor": "^4.3.1",
     "firebase-tools": "^10.7.0",
+    "resolve": "^1.22.0",
     "ts-proto": "^1.110.4",
     "wait-on": "^6.0.1"
   },
@@ -91,12 +93,8 @@
     "compile": "tsc",
     "format": "prettier --write --loglevel warn \"./src/**/*.ts{,x}\" --ignore-path \"./.custom-prettierignore\"",
     "proto": "./scripts/proto/get-proto.sh && ./scripts/proto/codegen.sh && prettier --write --loglevel warn \"./src/codec/**/*.ts\"",
-    "cy:open:local": "cypress open  --env configFile=local",
-    "cy:open:dev": "cypress open --env configFile=dev",
-    "cy:open:stage": "cypress open --env configFile=stage",
-    "cy:dev": "npx cypress run --browser chrome --env configFile=dev --spec 'cypress/integration/volumeTests/**.ts', 'cypress/integration/exchangeTokens.feature'",
-    "cy:stage": "npx cypress run --browser chrome  --env configFile=stage --spec 'cypress/integration/volumeTests/**.ts', 'cypress/integration/exchangeTokens.feature'",
-    "cy:local": "yarn start:local & wait-on http://localhost:3000 && cypress run --browser chrome --env configFile=local --spec 'cypress/integration/volumeTests/**.ts', 'cypress/integration/exchangeTokens.feature'"
+    "cy:open:local": "e2e open",
+    "cy:local": "yarn start:local & wait-on http://localhost:3000 && e2e run --browser chrome --spec 'e2e/integration/volumeTests/**.ts', 'e2e/integration/exchangeTokens.feature'"
   },
   "browserslist": [
     "last 2 chrome versions",
@@ -104,6 +102,6 @@
     "last 2 safari versions"
   ],
   "cypress-cucumber-preprocessor": {
-    "step_definitions": "cypress/steps-definition/"
+    "step_definitions": "e2e/steps-definition/"
   }
 }
